/docs/react/project-setup/react should not exist
