+++ homepage/homepage/app/docs/[framework]/[...slug]/page.tsx
@@ -1,13 +1,27 @@
 import fs from "fs";
 import path from "path";
 import { TableOfContents } from "@/components/docs/TableOfContents";
+import { Framework, frameworks, isValidFramework } from "@/lib/framework";
 import type { Toc } from "@stefanprobst/rehype-extract-toc";
 import { Prose } from "gcmp-design-system/src/app/components/molecules/Prose";
 
+function Error({ slugPath }: { slugPath: string }) {
+  return (
+    <Prose className="overflow-x-hidden lg:flex-1">
+      <h3>Error loading page: {slugPath}</h3>
+    </Prose>
+  );
+}
+
 export default async function Page({
-  params,
+  params: { slug, framework },
 }: { params: { slug: string[]; framework: string } }) {
-  const slugPath = params.slug.join("/");
+  const slugPath = slug.join("/");
+
+  // if the route ends in a framework name, return 404
+  // because we want the framework name to be in /docs/[framework]/[...slug]
+  if (isValidFramework(slug[slug.length - 1]))
+    return <Error slugPath={slugPath} />;
 
   try {
     let mdxSource;
@@ -15,7 +29,7 @@ export default async function Page({
       mdxSource = await import(`./${slugPath}.mdx`);
     } catch (error) {
       console.log("Error loading MDX file:" + slugPath, error);
-      mdxSource = await import(`./${slugPath}/${params.framework}.mdx`);
+      mdxSource = await import(`./${slugPath}/${framework}.mdx`);
     }
 
     const { default: Content, tableOfContents } = mdxSource;
@@ -30,11 +44,7 @@ export default async function Page({
     );
   } catch (error) {
     console.error("Error loading MDX file:" + slugPath, error);
-    return (
-      <Prose className="overflow-x-hidden lg:flex-1">
-        <h3>Error loading page: {slugPath}</h3>
-      </Prose>
-    );
+    return <Error slugPath={slugPath} />;
   }
 }
