+++ homepage/homepage/lib/framework.ts
@@ -1,19 +1,11 @@
-import { usePathname } from "next/navigation";
-
 export enum Framework {
   React = "react",
   ReactNative = "react-native",
   Vue = "vue",
 }
 
-function isValidFramework(value: string): value is Framework {
-  return Object.values(Framework).includes(value as Framework);
-}
+export const frameworks = Object.values(Framework);
 
-export const useFramework = () => {
-  const pathname = usePathname();
-  const framework = pathname.startsWith("/docs/")
-    ? pathname.split("/")[2]
-    : null;
-  return framework && isValidFramework(framework) ? framework : Framework.React;
-};
+export function isValidFramework(value: string): value is Framework {
+  return frameworks.includes(value as Framework);
+}
